Updated business list row item views

+++ data/src/main/java/com/vestrel00/business/search/data/entity/BusinessEntity.java
@@ -60,7 +60,7 @@ public abstract class BusinessEntity implements Entity {
 
     public abstract String price();
 
-    public abstract double rating();
+    public abstract float rating();
 
     public abstract boolean closed();
 
@@ -84,7 +84,7 @@ public abstract class BusinessEntity implements Entity {
 
         public abstract Builder price(String price);
 
-        public abstract Builder rating(double rating);
+        public abstract Builder rating(float rating);
 
         public abstract Builder closed(boolean closed);
